feat: Template refactor (Lead Asset) (Extreme WIP)
