--- packages/article-magazine-standard/src/article-lead-asset/crop-prop-types.js
@@ -1,8 +1,0 @@
-import PropTypes from "prop-types";
-
-const cropPropTypes = PropTypes.shape({
-  ratio: PropTypes.string.isRequired,
-  url: PropTypes.string.isRequired
-});
-
-export default cropPropTypes;
