+++ packages/article-skeleton/package.json
@@ -2,8 +2,8 @@
   "name": "@times-components/article-skeleton",
   "version": "0.1.2",
   "description": "The article skeleton",
-  "main": "dist/article-skeleton",
-  "dev": "src/article-skeleton",
+  "main": "dist",
+  "dev": "src",
   "scripts": {
     "fmt": "eslint . --fix && prettier --write '**/*.*'",
     "prettier:diff": "prettier --list-different '**/*.*'",
