+++ packages/article-skeleton/src/article-lead-asset/crop-prop-types.js
