+++ packages/article-skeleton/src/index.js
@@ -1,0 +1,2 @@
+export { default } from "./article-skeleton";
+export { default as LeadAsset } from "./article-lead-asset/article-lead-asset";
